refactor: make yajl_option to enum ParserOption
