--- crates/yajl/src/yajl_option.rs
@@ -1,6 +1,0 @@
-pub type yajl_option = libc::c_uint;
-pub const yajl_allow_partial_values: yajl_option = 16;
-pub const yajl_allow_multiple_values: yajl_option = 8;
-pub const yajl_allow_trailing_garbage: yajl_option = 4;
-pub const yajl_dont_validate_strings: yajl_option = 2;
-pub const yajl_allow_comments: yajl_option = 1;
