+++ examples/json-reformat/src/main.rs
@@ -2,8 +2,7 @@ use std::ptr::{addr_of, addr_of_mut};
 
 use ::libc;
 use yajl::{
-    parser::yajl_handle,
-    parser::{yajl_callbacks, Parser},
+    parser::{yajl_callbacks, yajl_handle, Parser},
     yajl_alloc::yajl_alloc_funcs,
     yajl_gen::{
         yajl_gen, yajl_gen_alloc, yajl_gen_array_close, yajl_gen_array_open, yajl_gen_beautify,
@@ -13,9 +12,8 @@ use yajl::{
         yajl_gen_string, yajl_gen_validate_utf8,
     },
     yajl_status::yajl_status,
-    yajl_tree::{
-        yajl_allow_comments, yajl_allow_multiple_values, yajl_dont_validate_strings, yajl_status_ok,
-    },
+    yajl_tree::yajl_status_ok,
+    ParserOption,
 };
 extern "C" {
 
@@ -155,7 +153,7 @@ unsafe fn main_0(argc: libc::c_int, argv: *mut *mut libc::c_char) -> libc::c_int
         g as *mut libc::c_void,
     );
     let parser = unsafe { &mut *hand };
-    parser.config(yajl_allow_comments, 1 as libc::c_int);
+    parser.config(ParserOption::AllowComments, true);
     while a < argc
         && *(*argv.offset(a as isize)).offset(0 as libc::c_int as isize) as libc::c_int
             == '-' as i32
@@ -168,11 +166,11 @@ unsafe fn main_0(argc: libc::c_int, argv: *mut *mut libc::c_char) -> libc::c_int
                     yajl_gen_config(g, yajl_gen_beautify, 0 as libc::c_int);
                 }
                 115 => {
-                    parser.config(yajl_allow_multiple_values, 1 as libc::c_int);
+                    parser.config(ParserOption::AllowMultipleValues, true);
                     STREAM_REFORMAT = 1 as libc::c_int;
                 }
                 117 => {
-                    parser.config(yajl_dont_validate_strings, 1 as libc::c_int);
+                    parser.config(ParserOption::DontValidateStrings, true);
                 }
                 101 => {
                     yajl_gen_config(g, yajl_gen_escape_solidus, 1 as libc::c_int);
